requirements.txt
peewee
pydantic
factory_boy
requests
funcy
loguru
python-telegram-bot
python-dotenv
PySocks
click